requirements.txt

sqlalchemy
pydantic==2.9.2
pydantic-core==2.23.4
pydantic-settings
passlib
python-jose
python-multipart
email-validator
alembic
psycopg2-binary
bcrypt
